requirements.txt
# This requirements file lists all third-party dependencies for this project.
#
# Run 'pip install -r requirements.txt -t lib/' to install these dependencies
# in `lib/` subdirectory.
#
# Note: The `lib` directory is added to `sys.path` by `appengine_config.py`.
Flask==0.10
pycrypto==2.6
GoogleAppEngineCloudStorageClient==0.6
requests==2.10.0
requests-toolbelt==0.6.2